adicionado método para código de produto, inicio da parte de ProdutoDAO

+++ src/ProjetoIntegrado2/src/Model/Produto.java
@@ -52,7 +52,7 @@ public class Produto {
         this.codProduto = codProduto;
     }
 
-    public double getPreco() {
+    public float getPreco() {
         return preco;
     }
